+++ CHANGELOG.md
@@ -48,6 +48,8 @@ Main (unreleased)
 
 - Add a description to Alloy DEB and RPM packages. (@rfratto)
 
+- Allow `pyroscope.scrape` to scrape `alloy.internal:12345`. (@hainenber)
+
 - The latest Windows Docker image is now pushed as `nanoserver-1809` instead of
   `latest-nanoserver-1809`. The old tag will no longer be updated, and will be
   removed in a future release. (@rfratto)